fix(scripts): Remove build.sh
build.sh removed due to CI testing on the multiple platforms, and local building
can be done with a single command.

This mostly stems from needing to set the linkers for the builds in .cargo/config, and the linker name is
different than what the ubuntu-latest runner is using (gcc).